+++ pipelex/cogt/llm/llm_models/llm_family.py
@@ -69,6 +69,7 @@ class LLMFamily(StrEnum):
     CLAUDE_3_5 = "claude-3.5"
     CLAUDE_3_7 = "claude-3.7"
     CLAUDE_4 = "claude-4"
+    CLAUDE_4_1 = "claude-4.1"
 
     MISTRAL_7B = "mistral-7b"
     MISTRAL_8X7B = "mistral-8x7b"
@@ -115,7 +116,7 @@ class LLMFamily(StrEnum):
                 | LLMFamily.GPT_5_CHAT
             ):
                 return LLMCreator.OPENAI
-            case LLMFamily.CLAUDE_3 | LLMFamily.CLAUDE_3_5 | LLMFamily.CLAUDE_3_7 | LLMFamily.CLAUDE_4:
+            case LLMFamily.CLAUDE_3 | LLMFamily.CLAUDE_3_5 | LLMFamily.CLAUDE_3_7 | LLMFamily.CLAUDE_4 | LLMFamily.CLAUDE_4_1:
                 return LLMCreator.ANTHROPIC
             case LLMFamily.BEDROCK_ANTHROPIC_CLAUDE:
                 return LLMCreator.ANTHROPIC